Refactor the Project and format the code accordingly...
