+++ demo/src/main/AndroidManifest.xml
@@ -30,12 +30,12 @@ Copyright 2024 Carizma Library All Rights Reserved.
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.CarizmaLibrary"
+        android:theme="@style/Theme.Sonux"
         tools:targetApi="31">
         <activity
             android:name=".activity.MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.CarizmaLibrary">
+            android:theme="@style/Theme.Sonux">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
