+++ demo/src/main/java/emmanuelmuturia/sonux/demo/activity/MainActivity.kt
@@ -1,20 +1,26 @@
 /*
- * Copyright 2024 Carizma Library
+ * Sonux  Copyright (C) 2024  Emmanuel Muturia™
+ * This program comes with ABSOLUTELY NO WARRANTY; for details type `show w'.
+ * This is free software, and you are welcome to redistribute it
+ * under certain conditions; type `show c' for details.
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * The hypothetical commands `show w' and `show c' should show the appropriate
+ * parts of the General Public License.  Of course, your program's commands
+ * might be different; for a GUI interface, you would use an "about box".
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * You should also get your employer (if you work as a programmer) or school,
+ * if any, to sign a "copyright disclaimer" for the program, if necessary.
+ * For more information on this, and how to apply and follow the GNU GPL, see
+ * <https://www.gnu.org/licenses/>.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-package emmanuelmuturia.sonux.activity
+ * The GNU General Public License does not permit incorporating your program
+ * into proprietary programs.  If your program is a subroutine library, you
+ * may consider it more useful to permit linking proprietary applications with
+ * the library.  If this is what you want to do, use the GNU Lesser General
+ * Public License instead of this License.  But first, please read
+ * <https://www.gnu.org/licenses/why-not-lgpl.html>.
+*/
+package emmanuelmuturia.sonux.demo.activity
 
 import android.content.Context
 import android.media.MediaPlayer
@@ -40,9 +46,9 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import emmanuelmuturia.sonux.theme.CarizmaLibraryTheme
-import emmanuelmuturia.sonux.R
-import emmanuelmuturia.sonux.data.extensions.to8D
+import emmanuelmuturia.sonux.demo.R
+import emmanuelmuturia.sonux.demo.theme.CarizmaLibraryTheme
+import emmanuelmuturia.sonux.library.extensions.to8D
 import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
